requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
django-filter==24.3
djangorestframework-simplejwt==5.5.0
psycopg[binary]==3.2.3 